added some routes to the main.dart file and createed the events screen
